+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Semantic versioning in our case means:
 - Forbids single element unpacking
 - Forbids to unpack lists with side-effects
 - Forbids to use miltiline strings except for assignments and docstrings
+- Forbids not returning anything in functions and methods starting with `get_`
 
 ### Bugfixes